fix: position bigint para evitar overflow ao criar lead
Date.now() (~1.7e12) excede o range de int4 (max ~2.1e9), causando 400
"Bad Request" do Supabase ao inserir leads.

- Migration 0001 atualizada para position bigint
- Nova migration 0002_position_bigint.sql faz ALTER em bancos existentes
- pgError(): mensagens de erro do Supabase agora mostram message + details
  + hint + code, em vez de cair no fallback genérico "Erro ao salvar."

+++ supabase/migrations/0001_init.sql
@@ -16,7 +16,7 @@ create table if not exists leads (
   estimated_value numeric(12,2),
   probability int check (probability is null or (probability between 0 and 100)),
   notes text,
-  position int not null default 0,
+  position bigint not null default 0,
   created_at timestamptz not null default now(),
   updated_at timestamptz not null default now()
 );
